split into different answer type into components
